Created dao sql server with MySQL DB Added OOP task company

+++ dao/src/main/java/com/stack/dao/entity/Album.java
@@ -20,9 +20,9 @@ import java.io.Serializable;
 @AllArgsConstructor
 public class Album implements Serializable {
     @JsonProperty("albumId")
-    private Long albumId;
+    private Integer albumId;
     @JsonProperty("title")
     private String title;
     @JsonProperty("artistId")
-    private Long artistId;
+    private Integer artistId;
 }
